Move Core Checklists into their own files
Core Checklists were cluttering up core.c, moving checklists to their own core_checklists.c file
